fix: merge navi-bar component into navi
